fix signed-overflow UB in 24 instruction headers (zpsfoperand)
GCC 13 aggressively eliminates sign-fixup code that follows
a signed integer overflow, exploiting the C++ rule that signed overflow
is undefined behavior.
GCC 11 does not. As a result, the same Spike source produces different simulation output on the two
platforms -- e.g. radd64 with rs1=0xEFFFFFFF10000000, rs2=0x8FFFFFFF70000000
gives x9=0xBFFFFFFF40000000 on Ubuntu 22 (correct) and 0x3FFFFFFF40000000
on Ubuntu 24 (broken).

Fix: perform intermediate arithmetic in a strictly wider type (__int128 /
__uint128_t for 64-bit operands; int64_t / int32_t / int16_t for narrower
lanes) and narrow back with an explicit cast. No UB anywhere, so GCC 11
and GCC 13 are forced to produce identical results.

Files changed (24):
  Group A (halving): radd64, rsub64, radd32, rsub32, radd16, rsub16,
    radd8, rsub8, rcras16, rcrsa16, rcras32, rcrsa32, uradd64, ursub64
  Group B (plain 64-bit add/sub): add64, sub64
  Group C (MAC/accumulate): smar64, smsr64, kmar64, kmsr64, smal,
    smalbb, smalbt, smaltt, smalda, smalxda, smslda, smslxda,
    smalds, smaldrs, smalxds

For kmar64/kmsr64: replaces the two sequential sat_add calls (which
required a 3-way branch to avoid intermediate saturation loss) with a
single __int128 accumulation followed by one clamp -- the ISA-mandated
"unlimited precision" semantic.

+++ riscv/insns/radd32.h
@@ -1,4 +1,4 @@
 require_rv64;
 P_LOOP(32, {
-  pd = ((int64_t)ps1 + ps2) >> 1;
+  pd = (int32_t)(((int64_t)ps1 + (int64_t)ps2) >> 1);
 })
