+++ riscv/insns/radd8.h
@@ -1,3 +1,4 @@
 P_LOOP(8, {
-  pd = (ps1 + ps2) >> 1;
+  // pd = (ps1 + ps2) >> 1;
+  pd = (int8_t)(((int16_t)ps1 + (int16_t)ps2) >> 1);
 })
