+++ riscv/insns/rcras16.h
@@ -1,5 +1,5 @@
 P_CROSS_LOOP(16, {
-  pd = (ps1 + ps2) >> 1;
+  pd = (int16_t)(((int32_t)ps1 + (int32_t)ps2) >> 1);
 }, {
-  pd = (ps1 - ps2) >> 1;
+  pd = (int16_t)(((int32_t)ps1 - (int32_t)ps2) >> 1);
 })
