+++ riscv/insns/rcrsa32.h
@@ -1,6 +1,6 @@
 require_rv64;
 P_CROSS_LOOP(32, {
-  pd = ((uint64_t)ps1 - ps2) >> 1;
+  pd = (int32_t)(((int64_t)ps1 - (int64_t)ps2) >> 1);
 }, {
-  pd = ((uint64_t)ps1 + ps2) >> 1;
+  pd = (int32_t)(((int64_t)ps1 + (int64_t)ps2) >> 1);
 })
