+++ riscv/insns/rsub16.h
@@ -1,3 +1,4 @@
 P_LOOP(16, {
-  pd = (ps1 - ps2) >> 1;
+  // pd = (ps1 - ps2) >> 1;
+  pd = (int16_t)(((int32_t)ps1 - (int32_t)ps2) >> 1);
 })
